+++ packages/desktop/src/index.ts
@@ -21,6 +21,7 @@ export * from './header'
 export * from './drop-down'
 export * from './phone-field'
 export * from './status'
+export * from './logo-link'
 
 export {TextField as MaskTextField} from './text-field'
 export {PasswordField as MaskPasswordField} from './password-field'