+++ packages/mobile/src/index.ts
@@ -18,6 +18,7 @@ export * from './header'
 export * from './menu'
 export * from './pagination'
 export * from './status'
+export * from './logo-link'
 
 export {TextField as MaskTextField} from './text-field'
 export {PasswordField as MaskPasswordField} from './password-field'